Remove func.rs from nanobind backend
